fix: Implement critical production blockers

+++ q_mini_wasm_v2/core/moe/gf3_layers.hpp
@@ -133,8 +133,9 @@ public:
 
     /**
      * @brief Get sparsity (fraction of zero weights)
+     * @return Q24.8 fixed point representation
      */
-    float GetSparsity() const;
+    uint32_t GetSparsity() const;
 
 private:
     LayerConfig config_;
